ref: rename multiexpr to wideexpr
